requirements.txt
amqp==2.5.0
anyjson==0.3.3
attrs==19.1.0
Babel==2.7.0
backports-abc==0.5
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==1.11.21
django-celery==3.3.0
django-filter==2.0.0
django-pkgconf==0.3.0
django-redis==4.10.0
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
flower==0.9.3
futures==3.2.0
idna==2.8
kombu==4.6.1
Markdown==3.1.1
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.1
redis==3.2.1
requests==2.22.0
singledispatch==3.4.0.3
six==1.12.0
tornado==5.1.1
urllib3==1.25.3
vine==1.3.0
